* Use os.path.isfile instead of try/catch

+++ pokemongo_bot/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+import os
 import logging
 import googlemaps
 import json
@@ -55,8 +56,8 @@ class PokemonGoBot(object):
                 key=
                 lambda x: distance(self.position[0], self.position[1], x['latitude'], x['longitude']))
 
-            try:
-                user_web_catchable = 'web/catchable-%s.json' % (self.config.username)
+            user_web_catchable = 'web/catchable-%s.json' % (self.config.username)
+            if os.path.isfile(user_web_catchable): # only write to file if it exists
                 for pokemon in cell['catchable_pokemons']:
                     with open(user_web_catchable, 'w') as outfile:
                         json.dump(pokemon, outfile)
@@ -65,9 +66,6 @@ class PokemonGoBot(object):
                         break
                     with open(user_web_catchable, 'w') as outfile:
                         json.dump({}, outfile)
-            except:
-                # web folder probably doesn't exist... just skip
-                pass
 
         if (self.config.mode == "all" or self.config.mode == "poke"
             ) and 'wild_pokemons' in cell and len(cell['wild_pokemons']) > 0:
@@ -239,11 +237,10 @@ class PokemonGoBot(object):
             'inventory_delta']['inventory_items']
 
         user_web_inventory = 'web/inventory-%s.json' % (self.config.username)
-        try:
+        if os.path.isfile(user_web_inventory):
             with open(user_web_inventory, 'w') as outfile:
                 json.dump(inventory_dict, outfile)
-        except:
-            pass
+
         # get player balls stock
         # ----------------------
         balls_stock = {1: 0, 2: 0, 3: 0, 4: 0}
