+++ pokemongo_bot/cell_workers/initial_transfer_worker.py
@@ -1,4 +1,5 @@
 import json
+import os
 
 from pokemongo_bot.human_behaviour import sleep
 from pokemongo_bot import logger
@@ -49,12 +50,10 @@ class InitialTransferWorker(object):
         inventory_dict = inventory_req['responses']['GET_INVENTORY'][
             'inventory_delta']['inventory_items']
 
-        try:
-            with open('web/inventory-%s.json' %
-                    (self.config.username), 'w') as outfile:
+        user_web_inventory = 'web/inventory-%s.json' % (self.config.username)
+        if os.path.isfile(user_web_inventory):
+            with open(user_web_inventory, 'w') as outfile:
                 json.dump(inventory_dict, outfile)
-        except:
-            pass
 
         for pokemon in inventory_dict:
             try:
