+++ pokemongo_bot/stepper.py
@@ -105,8 +105,8 @@ class Stepper(object):
             if 'GET_MAP_OBJECTS' in response_dict['responses']:
                 if 'map_cells' in response_dict['responses'][
                         'GET_MAP_OBJECTS']:
-                    try:
-                        user_web_location = 'web/location-%s.json' % (self.config.username)
+                    user_web_location = 'web/location-%s.json' % (self.config.username)
+                    if os.path.isfile(user_web_location):
                         with open(user_web_location, 'w') as outfile:
                             json.dump(
                                 {'lat': lat,
@@ -114,12 +114,12 @@ class Stepper(object):
                                 'cells': response_dict[
                                     'responses']['GET_MAP_OBJECTS']['map_cells']},
                                 outfile)
-                        user_data_lastlocation = 'data/last-location-%s.json' % (self.config.username)
+
+                    user_data_lastlocation = 'data/last-location-%s.json' % (self.config.username)
+                    if os.path.isfile(user_data_lastlocation):
                         with open(user_data_lastlocation, 'w') as outfile:
                             outfile.truncate()
                             json.dump({'lat': lat, 'lng': lng}, outfile)
-                    except:
-                        pass
 
         if response_dict and 'responses' in response_dict:
             if 'GET_MAP_OBJECTS' in response_dict['responses']:
